fix(cron): move parseDurationToMs into lib/duration so the web build passes
Next.js 15 only allows a fixed set of named exports from `route.ts` files
(HTTP method handlers plus config like `dynamic`, `runtime`, etc.). The
cron jobs route was also exporting `parseDurationToMs` so the test could
import it, which now fails the production build:

  Type error: Route "app/api/cron/jobs/route.ts" does not match the
  required types of a Next.js Route.
    "parseDurationToMs" is not a valid Route export field.

Extracted the helper into a small shared `apps/web/lib/duration.ts`
module, switched the route and the test to import it from there. No
behavioural change; all 16 cron tests still pass and `pnpm web:build`
is green again.

Made-with: Cursor
